+++ gulp/src/common/create-redux-store.js
@@ -19,6 +19,4 @@ export default function createReduxStore(reducer, defaultState, thunkExtra) {
     applyMiddleware(thunkMiddleware.withExtraArgument(thunkExtra)),
     applyMiddleware(promiseMiddleware),
     devToolsMiddleware));
-    // redux devtools
-    //window.devToolsExtension ? window.devToolsExtension() : f => f));
 }
