+++ gulp/src/common/id-generator.js
