+++ gulp/src/reporting/spec/reportEngineSpec.js
@@ -214,164 +214,7 @@ describe('ReportConfiguration', () => {
   it('can visit callbacks to push initial state', () => {
     config.runCallbacks();
     expect(fakeComponent.onNameChanged).toHaveBeenCalledWith("defaultName");
-    expect(fakeComponent.onUpdateFilter).toHaveBeenCalledWith({});
     expect(fakeComponent.onErrorsChanged).toHaveBeenCalledWith({});
-  });
-
-  it('can set simple filters', () => {
-    config.setFilter('city', 'Indianapolis');
-    expect(config.getFilter()).toEqual({city: 'Indianapolis'});
-    expect(fakeComponent.onUpdateFilter).toHaveBeenCalledWith({
-      city: 'Indianapolis',
-    });
-    config.setFilter('state', 'Indiana');
-    expect(config.getFilter()).toEqual({
-      city: 'Indianapolis',
-      state: 'Indiana',
-    });
-    expect(fakeComponent.onUpdateFilter).toHaveBeenCalledWith({
-      city: 'Indianapolis',
-      state: 'Indiana',
-    });
-  });
-
-  it('can set simple filters to object values', () => {
-    config.setFilter('locations', {'city': 'Indianapolis'});
-    expect(config.getFilter()).toEqual({'locations': {city: 'Indianapolis'}});
-  });
-
-  it('can set simple filters to date range values', () => {
-    config.setFilter('date_range', ['2016-01-01', '2016-01-02']);
-    expect(config.getFilter()).toEqual(
-      {'date_range': ['2016-01-01', '2016-01-02']});
-  });
-
-  it('treats null filter values as removal', () => {
-    config.setFilter('city', 'Indianapolis');
-    config.setFilter('city', null);
-    expect(config.getFilter()).toEqual({});
-    expect(fakeComponent.onUpdateFilter).toHaveBeenCalledWith({});
-  });
-
-  it('treats undefined filter values as removal', () => {
-    config.setFilter('city', 'Indianapolis');
-    config.setFilter('city', undefined);
-    expect(config.getFilter()).toEqual({});
-    expect(fakeComponent.onUpdateFilter).toHaveBeenCalledWith({});
-  });
-
-  it('can set changes to multifilters', () => {
-    config.addToMultifilter('tag', {value: 'blue', display: 'Blue'});
-    expect(config.getFilter()).toEqual({tag: ['blue']});
-    expect(fakeComponent.onUpdateFilter).toHaveBeenCalledWith({
-      'tag': [{value: 'blue', display: 'Blue'}],
-    });
-  });
-
-  it("won't add dups in multifilters", () => {
-    config.addToMultifilter('tag', {value: 'blue', display: 'Blue'});
-    config.addToMultifilter('tag', {value: 'blue', display: 'Blue'});
-    expect(config.getFilter()).toEqual({tag: ['blue']});
-    expect(fakeComponent.onUpdateFilter).toHaveBeenCalledWith({
-      'tag': [{value: 'blue', display: 'Blue'}],
-    });
-  });
-
-  it('can remove items from multifilters', () => {
-    config.addToMultifilter('tag', {value: 'red', display: 'Red'});
-    expect(fakeComponent.onUpdateFilter).toHaveBeenCalledWith({
-      'tag': [{value: 'red', display: 'Red'}],
-    });
-    config.addToMultifilter('tag', {value: 'blue', display: 'Blue'});
-    expect(fakeComponent.onUpdateFilter).toHaveBeenCalledWith({
-      'tag': [
-        {value: 'red', display: 'Red'},
-        {value: 'blue', display: 'Blue'},
-      ],
-    });
-    config.removeFromMultifilter('tag', {value: 'red'});
-    expect(fakeComponent.onUpdateFilter).toHaveBeenCalledWith({
-      'tag': [{value: 'blue', display: 'Blue'}],
-    });
-    config.removeFromMultifilter('tag', {value: 'red'});
-    expect(config.getFilter()).toEqual({tag: ['blue']});
-    expect(fakeComponent.onUpdateFilter).toHaveBeenCalledWith({
-      'tag': [{value: 'blue', display: 'Blue'}],
-    });
-    config.removeFromMultifilter('tag', {value: 'blue'});
-    expect(config.getFilter()).toEqual({});
-    expect(fakeComponent.onUpdateFilter).toHaveBeenCalledWith({});
-  });
-
-  it('can remember and/or filters', () => {
-    config.addToAndOrFilter('tag', 0, {value: 'red', display: 'Red'});
-    expect(config.getFilter()).toEqual({tag: [['red']]});
-    expect(fakeComponent.onUpdateFilter).toHaveBeenCalledWith({
-      'tag': [[{value: 'red', display: 'Red'}]],
-    });
-  });
-
-  it('can remove items from and/or filters', () => {
-    config.addToAndOrFilter('tag', 0, {value: 'red', display: 'Red'});
-    expect(fakeComponent.onUpdateFilter).toHaveBeenCalledWith({
-      'tag': [[{value: 'red', display: 'Red'}]],
-    });
-    config.addToAndOrFilter('tag', 0, {value: 'blue', display: 'Blue'});
-    expect(fakeComponent.onUpdateFilter).toHaveBeenCalledWith({
-      'tag': [[
-        {value: 'red', display: 'Red'},
-        {value: 'blue', display: 'Blue'},
-      ]],
-    });
-    config.removeFromAndOrFilter('tag', 0, {value: 'red'});
-    expect(fakeComponent.onUpdateFilter).toHaveBeenCalledWith({
-      'tag': [[{value: 'blue', display: 'Blue'}]],
-    });
-    config.removeFromAndOrFilter('tag', 0, {value: 'red'});
-    expect(fakeComponent.onUpdateFilter).toHaveBeenCalledWith({
-      'tag': [[{value: 'blue', display: 'Blue'}]],
-    });
-    expect(config.getFilter()).toEqual({
-      tag: [['blue']],
-    });
-    config.removeFromAndOrFilter('tag', 0, {value: 'blue'});
-    expect(fakeComponent.onUpdateFilter).toHaveBeenCalledWith({});
-    expect(config.getFilter()).toEqual({});
-  });
-
-  it('removes empty tag lists on demand for and/or filters', () => {
-    config.addToAndOrFilter('tag', 0, {value: 'red', display: 'Red'});
-    expect(fakeComponent.onUpdateFilter).toHaveBeenCalledWith({
-      'tag': [[{value: 'red', display: 'Red'}]],
-    });
-    config.addToAndOrFilter('tag', 1, {value: 'red', display: 'Red'});
-    expect(fakeComponent.onUpdateFilter).toHaveBeenCalledWith({
-      'tag': [
-        [{value: 'red', display: 'Red'}],
-        [{value: 'red', display: 'Red'}],
-      ],
-    });
-    config.addToAndOrFilter('tag', 0, {value: 'blue', display: 'Blue'});
-    expect(fakeComponent.onUpdateFilter).toHaveBeenCalledWith({
-      'tag': [
-        [{value: 'red', display: 'Red'}, {value: 'blue', display: 'Blue'}],
-        [{value: 'red', display: 'Red'}],
-      ],
-    });
-    config.removeFromAndOrFilter('tag', 0, {value: 'red'});
-    expect(fakeComponent.onUpdateFilter).toHaveBeenCalledWith({
-      'tag': [
-        [{value: 'blue', display: 'Blue'}],
-        [{value: 'red', display: 'Red'}],
-      ],
-    });
-    config.removeFromAndOrFilter('tag', 0, {value: 'blue'});
-    expect(fakeComponent.onUpdateFilter).toHaveBeenCalledWith({
-      'tag': [
-        [{value: 'red', display: 'Red'}],
-      ],
-    });
-    expect(config.getFilter()).toEqual({tag: [['red']]});
   });
 
   it('can run the report', promiseTest(async () => {
